Modify in response to reviewer feedback
The reviewers wanted two changes to Figures 2 and 3.

They wanted Figures 2B and 3B to have the same axes. I did this by
swapping out the `facet_wrap()` in panel 2 to instead use different
figures, so I could specify the `ylim()` separately for each. Then,
I used the patchwork `plot_layout(design = design)` specification to lay
out the relative placement of the 3 plots.

The other desired change was to add which cell line was sensitive and
which was resistant into the legend.